Makes FlashHash Enumerable
This allows the iteration over the flash messages in a view, like Rails
does.

Added a method to clear the hash cache, so that by clearing the session
and the cache we can effectivly empty the flash messages hash.

Also added a Gemfile with the test dependencies and corrected the tests
(except the integration test). This will make it easier for the next
person trying to contribute.